tests: split massive dashboard/tests.py file into smaller ones
